Change return value type of structureChangedEventHandler_queryInterface
error before this commit:
```
C:\Users\hnakamur\w32uiautomation\examples\runcalc>go run main.go
panic: compileCallback: expected function with one uintptr-sized result
goroutine 1 [running]:
syscall.compileCallback({0x3c8f80?, 0x3e7078?}, 0x0?)
        c:/go/src/runtime/syscall_windows.go:273 +0x43e
syscall.NewCallback(...)
        c:/go/src/syscall/syscall_windows.go:181
github.com/hnakamur/w32uiautomation.NewStructureChangedEventHandler(0x3ba49b?)
        C:/Users/hnakamur/w32uiautomation/iuiautomationstructurechangedeventhandler.go:88 +0x35
github.com/hnakamur/w32uiautomation.waitChildAdded(0x0?, 0x0?)
        C:/Users/hnakamur/w32uiautomation/waitfind.go:26 +0x3d
github.com/hnakamur/w32uiautomation.WaitFindFirst(0x8?, 0x23beeb65998?, 0x0?, 0x0?)
        C:/Users/hnakamur/w32uiautomation/waitfind.go:20 +0x39
main.runCalc()
        C:/Users/hnakamur/w32uiautomation/examples/runcalc/main.go:42 +0x145
main.main()
        C:/Users/hnakamur/w32uiautomation/examples/runcalc/main.go:100 +0x45
exit status 2
```

error after this commit:
```
C:\Users\hnakamur\w32uiautomation\examples\runcalc>go run main.go
panic: compileCallback: expected function with one uintptr-sized result
goroutine 1 [running]:
syscall.compileCallback({0xfb6d20?, 0xfd7070?}, 0x0?)
        c:/go/src/runtime/syscall_windows.go:273 +0x43e
syscall.NewCallback(...)
        c:/go/src/syscall/syscall_windows.go:181
github.com/hnakamur/w32uiautomation.NewStructureChangedEventHandler(0xfaa49b?)
        C:/Users/hnakamur/w32uiautomation/iuiautomationstructurechangedeventhandler.go:89 +0x52
github.com/hnakamur/w32uiautomation.waitChildAdded(0x0?, 0x0?)
        C:/Users/hnakamur/w32uiautomation/waitfind.go:26 +0x3d
github.com/hnakamur/w32uiautomation.WaitFindFirst(0x8?, 0x1953ae36988?, 0x0?, 0x0?)
        C:/Users/hnakamur/w32uiautomation/waitfind.go:20 +0x39
main.runCalc()
        C:/Users/hnakamur/w32uiautomation/examples/runcalc/main.go:42 +0x145
main.main()
        C:/Users/hnakamur/w32uiautomation/examples/runcalc/main.go:100 +0x45
exit status 2
```

+++ iuiautomationstructurechangedeventhandler.go
@@ -54,7 +54,7 @@ func (h *IUIAutomationStructureChangedEventHandler) VTable() *IUIAutomationStruc
 	return (*IUIAutomationStructureChangedEventHandlerVtbl)(unsafe.Pointer(h.RawVTable))
 }
 
-func structureChangedEventHandler_queryInterface(this *ole.IUnknown, iid *ole.GUID, punk **ole.IUnknown) uint32 {
+func structureChangedEventHandler_queryInterface(this *ole.IUnknown, iid *ole.GUID, punk **ole.IUnknown) uintptr {
 	*punk = nil
 	if ole.IsEqualGUID(iid, ole.IID_IUnknown) ||
 		ole.IsEqualGUID(iid, ole.IID_IDispatch) {
